разделяет код на модули
